// projeto/web/angular_p2/src/api/generated/models/getApiOrdersParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */

export type GetApiOrdersParams = {
  /**
   * Campo de ordenacao (apenas 1).
   */
  'sort-field'?: string;
  /**
   * Direcao da ordenacao: asc|desc.
   */
  'sort-order'?: string;
  /**
   * Pagina (default 1).
   */
  page?: number;
  /**
   * Limite por pagina (default 25).
   */
  limit?: number;
};