// projeto/web/angular_p2/src/api/generated/models/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { EOrderStatus } from './eOrderStatus';
import type { OrderItem } from './orderItem';

export interface Order {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  tableNumber?: number | string;
  status?: EOrderStatus;
  /** @nullable */
  name?: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  userId?: number | string;
  orderItems?: OrderItem[];
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  createdAt?: string;
  updatedAt?: string;
}